// src/app/demo/models/modelsgeneral/tasacambioviewmodel.ts
export class TasaCambio {
    taca_Id?: number;
    mone_A?: string;
    mone_B?: string;
    taca_ValorA?: number;
    taca_ValorB?: number;
    moneda_A?: string;
    moneda_B?: string;
    usua_Creacion?: number;
    usua_Modificacion?: number;
    usuaCreacion?: string;
    usuaModificacion?: string;
    taca_FechaCreacion?: string;
    taca_FechaModificacion?: string;    

    codigo?: string;
}

export class MonedaA{
    mone_A?: string;
    moneda_A?: string;
}

export class MonedaB{
    mone_B?: string;
    moneda_B?: string;
}