// src/app/demo/models/modelsgeneral/monedaviewmodel.ts
export class Moneda{
    codigo?: number;
    mone_Id?: number;
    mone_Nombre?:string;
    mone_Abreviatura?:string;
    pais_Id?:number;
    pais_Nombre?:string;
    usua_Creacion?:number;
    usua_Modificacion?: number;
    usuaCreacion?:string;
    usuaModificacion?:string;
    mone_FechaCreacion?: string;
    mone_FechaModificacion?: string;
    mone_Estado?: boolean;

}